clean up + fix bug with empty repositories
lots of little style errors, and some not-so-little semantic
issues. most importantly, previous behavior was to create the archive
file in preparation for writing the download but if the archive did
not download (e.g. because the repo is empty) the file was left on
disk with no data. now, the file will be deleted if no data could be written.